requirements.txt
Django==4.1.5
requests==2.28.1
python-dotenv==0.20.0

